refactor(exporter-shopware-6) Rename CustomField mapper
