+++ module/exporter-shopware-6/src/Infrastructure/Builder/CustomFieldBuilder.php
@@ -12,17 +12,17 @@ use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 
-class Shopware6CustomFieldBuilder
+class CustomFieldBuilder
 {
     /**
-     * @var Shopware6CustomFieldMapperInterface[]
+     * @var CustomFieldMapperInterface[]
      */
     private array $collection;
 
-    public function __construct(Shopware6CustomFieldMapperInterface ...$collection)
+    public function __construct(CustomFieldMapperInterface ...$collection)
     {
         $this->collection = $collection;
     }
