+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigMediaMapper.php
@@ -9,18 +9,18 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\AbstractNumericAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\AbstractImageAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
-use Ergonode\ExporterShopware6\Infrastructure\Model\Basic\Shopware6CustomFieldConfig;
 
-class Shopware6CustomFieldConfigNumericMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigMediaMapper implements CustomFieldMapperInterface
 {
-    private const TYPE = 'number';
-    private const NUMBER_TYPE = 'float';
+    private const TYPE = 'text';
+    private const CUSTOM_FIELD_TYPE = 'media';
+    private const COMPONENT_NAME = 'sw-media-field';
 
     public function map(
         Shopware6Channel $channel,
@@ -30,13 +30,10 @@ class Shopware6CustomFieldConfigNumericMapper implements Shopware6CustomFieldMap
         ?Language $language = null
     ): AbstractShopware6CustomField {
 
-        if ($attribute->getType() === AbstractNumericAttribute::TYPE) {
+        if ($attribute->getType() === AbstractImageAttribute::TYPE) {
             $shopware6CustomField->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setCustomFieldType(self::TYPE);
-            if ($shopware6CustomField->getConfig() instanceof Shopware6CustomFieldConfig) {
-                $shopware6CustomField->getConfig()->setNumberType(self::NUMBER_TYPE);
-            }
+            $shopware6CustomField->getConfig()->setCustomFieldType(self::CUSTOM_FIELD_TYPE);
+            $shopware6CustomField->getConfig()->setComponentName(self::COMPONENT_NAME);
         }
 
         return $shopware6CustomField;
