+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigMultiSelectMapper.php
@@ -9,18 +9,17 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\AbstractTextareaAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\MultiSelectAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 
-class Shopware6CustomFieldConfigTextareaMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigMultiSelectMapper implements CustomFieldMapperInterface
 {
-    private const TYPE = 'text';
-    private const CUSTOM_FIELD_TYPE = 'textEditor';
-    private const COMPONENT_NAME = 'sw-text-editor';
+    private const TYPE = 'select';
+    private const COMPONENT_NAME = 'sw-multi-select';
 
     public function map(
         Shopware6Channel $channel,
@@ -30,10 +29,10 @@ class Shopware6CustomFieldConfigTextareaMapper implements Shopware6CustomFieldMa
         ?Language $language = null
     ): AbstractShopware6CustomField {
 
-        if ($attribute->getType() === AbstractTextareaAttribute::TYPE) {
+        if ($attribute->getType() === MultiSelectAttribute::TYPE) {
             $shopware6CustomField->setType(self::TYPE);
             $shopware6CustomField->getConfig()->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setCustomFieldType(self::CUSTOM_FIELD_TYPE);
+            $shopware6CustomField->getConfig()->setCustomFieldType(self::TYPE);
             $shopware6CustomField->getConfig()->setComponentName(self::COMPONENT_NAME);
         }
