+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigOptionsMapper.php
@@ -14,11 +14,11 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Domain\Repository\LanguageRepositoryInterface;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Basic\Shopware6CustomFieldConfig;
 
-class Shopware6CustomFieldConfigOptionsMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigOptionsMapper implements CustomFieldMapperInterface
 {
     private OptionQueryInterface $optionQuery;
