+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigSelectMapper.php
@@ -9,18 +9,17 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\AbstractDateAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\SelectAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
-use Ergonode\ExporterShopware6\Infrastructure\Model\Basic\Shopware6CustomFieldConfig;
 
-class Shopware6CustomFieldConfigDateMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigSelectMapper implements CustomFieldMapperInterface
 {
-    private const TYPE = 'datetime';
-    private const CONFIG_TYPE = 'date';
+    private const TYPE = 'select';
+    private const COMPONENT_NAME = 'sw-single-select';
 
     public function map(
         Shopware6Channel $channel,
@@ -30,13 +29,11 @@ class Shopware6CustomFieldConfigDateMapper implements Shopware6CustomFieldMapper
         ?Language $language = null
     ): AbstractShopware6CustomField {
 
-        if ($attribute->getType() === AbstractDateAttribute::TYPE) {
+        if ($attribute->getType() === SelectAttribute::TYPE) {
             $shopware6CustomField->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setType(self::CONFIG_TYPE);
-            $shopware6CustomField->getConfig()->setCustomFieldType(self::CONFIG_TYPE);
-            if ($shopware6CustomField->getConfig() instanceof Shopware6CustomFieldConfig) {
-                $shopware6CustomField->getConfig()->setDateType(self::TYPE);
-            }
+            $shopware6CustomField->getConfig()->setType(self::TYPE);
+            $shopware6CustomField->getConfig()->setCustomFieldType(self::TYPE);
+            $shopware6CustomField->getConfig()->setComponentName(self::COMPONENT_NAME);
         }
 
         return $shopware6CustomField;
