+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigTextareaMapper.php
@@ -9,17 +9,18 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\MultiSelectAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\AbstractTextareaAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 
-class Shopware6CustomFieldConfigMultiSelectMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigTextareaMapper implements CustomFieldMapperInterface
 {
-    private const TYPE = 'select';
-    private const COMPONENT_NAME = 'sw-multi-select';
+    private const TYPE = 'text';
+    private const CUSTOM_FIELD_TYPE = 'textEditor';
+    private const COMPONENT_NAME = 'sw-text-editor';
 
     public function map(
         Shopware6Channel $channel,
@@ -29,10 +30,10 @@ class Shopware6CustomFieldConfigMultiSelectMapper implements Shopware6CustomFiel
         ?Language $language = null
     ): AbstractShopware6CustomField {
 
-        if ($attribute->getType() === MultiSelectAttribute::TYPE) {
+        if ($attribute->getType() === AbstractTextareaAttribute::TYPE) {
             $shopware6CustomField->setType(self::TYPE);
             $shopware6CustomField->getConfig()->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setCustomFieldType(self::TYPE);
+            $shopware6CustomField->getConfig()->setCustomFieldType(self::CUSTOM_FIELD_TYPE);
             $shopware6CustomField->getConfig()->setComponentName(self::COMPONENT_NAME);
         }
