+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldConfigUnitMapper.php
@@ -13,11 +13,11 @@ use Ergonode\Attribute\Domain\Entity\Attribute\AbstractUnitAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Basic\Shopware6CustomFieldConfig;
 
-class Shopware6CustomFieldConfigUnitMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldConfigUnitMapper implements CustomFieldMapperInterface
 {
     private const TYPE = 'number';
     private const NUMBER_TYPE = 'float';
