+++ module/exporter-shopware-6/src/Infrastructure/Mapper/CustomField/CustomFieldNameMapper.php
@@ -9,19 +9,14 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\AbstractImageAttribute;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CustomFieldMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CustomFieldMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
 
-class Shopware6CustomFieldConfigMediaMapper implements Shopware6CustomFieldMapperInterface
+class CustomFieldNameMapper implements CustomFieldMapperInterface
 {
-    private const TYPE = 'text';
-    private const CUSTOM_FIELD_TYPE = 'media';
-    private const COMPONENT_NAME = 'sw-media-field';
-
     public function map(
         Shopware6Channel $channel,
         Export $export,
@@ -29,12 +24,7 @@ class Shopware6CustomFieldConfigMediaMapper implements Shopware6CustomFieldMappe
         AbstractAttribute $attribute,
         ?Language $language = null
     ): AbstractShopware6CustomField {
-
-        if ($attribute->getType() === AbstractImageAttribute::TYPE) {
-            $shopware6CustomField->setType(self::TYPE);
-            $shopware6CustomField->getConfig()->setCustomFieldType(self::CUSTOM_FIELD_TYPE);
-            $shopware6CustomField->getConfig()->setComponentName(self::COMPONENT_NAME);
-        }
+        $shopware6CustomField->setName($attribute->getCode()->getValue());
 
         return $shopware6CustomField;
     }
