--- module/exporter-shopware-6/src/Infrastructure/Mapper/Shopware6CustomFieldMapperInterface.php
@@ -1,26 +1,0 @@
-<?php
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE.txt for license details.
- */
-
-declare(strict_types=1);
-
-namespace Ergonode\ExporterShopware6\Infrastructure\Mapper;
-
-use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Core\Domain\ValueObject\Language;
-use Ergonode\Channel\Domain\Entity\Export;
-use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Model\AbstractShopware6CustomField;
-
-interface Shopware6CustomFieldMapperInterface
-{
-    public function map(
-        Shopware6Channel $channel,
-        Export $export,
-        AbstractShopware6CustomField $shopware6CustomField,
-        AbstractAttribute $attribute,
-        ?Language $language = null
-    ): AbstractShopware6CustomField;
-}
